refactor(solver): extract per-query eval budget into query_budget module
Keep evaluate.rs under the 2000-line ceiling by moving the cross-instance
per-query operation budget (thread-locals, EvalQueryFrame, the resolver, and its
unit tests) into a dedicated evaluation/evaluate/query_budget.rs submodule, and
expose a single TypeEvaluator::enter_eval_query_budget entry point. No behaviour
change.
